Completely change nearly everything
Well I now have a of better idea of how this should be structured. Oh
and this one compiles and runs, try running any Legends of Equestria
client and a login server (either mlkj's, old [LoEWCT] or LoE Movie
Maker, but with the game server disabled) and try connecting to this
server, the server will print out user info upon connection. That is all
it does at the moment really. Everything else here are still barebones
stuff.